requirements.txt
django
gunicorn
python-dotenv
whitenoise
django-admin-soft-dashboard==1.0.11
djongo==1.3.6
pymongo==4.10.1
sqlparse==0.2.4
pillow
xhtml2pdf
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz
